Allow disabling either IPv4 or IPv6.
This is particularly useful in situations where the underlying network doesn’t support IPv6. In particular its possible when running in docker that the OS allows setting up the IPv6 socket but when you go to send on it, then it will fail.

Also the test was refactored to fix some linter warnings about calling t.Fatalf from within a non-test go routine. Now that go routine will emit an error from a chan and the main test routine receives from it to determine whether the test should pass or fail.